Refactor Kademlia implementation into modular structure
 This commit overhauls the Kademlia DHT node implementation to adopt a more modular structure. The changes include:

 - Extracting the `KBucket` logic into its own file `kbucket.rs`, encapsulating the k-bucket functionality and
 improving code maintainability.
 - Moving the main node logic to `node.rs`, which now contains the `KademliaNode` struct and associated methods for
 node operations such as bootstrapping, message handling, and running the main event loop.
 - Creating `routing_table.rs` to handle the routing table logic separately, allowing for clearer separation of
 concerns and easier future enhancements to the routing mechanism.
 - Defining message types and results in `message.rs`, which simplifies the message handling process and makes the
 message format more explicit.
 - Introducing `utils.rs` to store common constants and utility functions, such as the `NodeId` struct and its
 associated methods, which are used across different modules.
